+++ packages/react-responsive-values/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/react-responsive-values'
+export * from './types'
